+++ cmake/third-party.cmake
@@ -37,7 +37,7 @@ GIT_SHALLOW ON)
 # Core feature providers
 FetchContent_Declare( slang
   GIT_REPOSITORY https://github.com/MikePopoloski/slang.git
-  GIT_TAG v7.0
+  GIT_TAG v8.0
   GIT_SHALLOW ON)